requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-generator==1.10
attrs==25.4.0
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dj-database-url==2.3.0
Django==4.2.27
djangorestframework==3.16.1
fake-useragent==2.2.0
gunicorn==25.0.1
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
kombu==5.6.2
lxml==6.0.2
Markdown==3.10.1
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==26.0
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
selenium==4.40.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
sqlparse==0.5.5
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.5.3
websocket-client==1.9.0
whitenoise==6.11.0
wsproto==1.3.2
zipp==3.23.0
